moved sim to robotContainer
