+++ src/main/java/frc/robot/subsystems/shooter/ShooterIO.java
@@ -2,7 +2,6 @@ package frc.robot.subsystems.shooter;
 
 import static edu.wpi.first.units.Units.*;
 
-import org.ironmaple.simulation.drivesims.SwerveDriveSimulation;
 import org.littletonrobotics.junction.AutoLog;
 import edu.wpi.first.units.measure.*;
 
@@ -28,5 +27,4 @@ public interface ShooterIO {
 	public default void stop() {}
 	public default void setWheelVelocitySetpoint(AngularVelocity topVelocity, AngularVelocity bottomVelocity) {}
 	public default void setWheelVoltage(Voltage volts) {}
-	public default void launchAlgae(SwerveDriveSimulation swerve){}
 }
